requirements.txt
appdirs==1.4.4
appnope==0.1.4
argcomplete==3.2.2
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli @ file:///private/tmp/python-brotli-20231117-5715-qv008m/Brotli-1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi @ file:///private/tmp/python-certifi-20231118-5270-132s65u/certifi-2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.2
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msgpack==1.0.8
mutagen @ file:///private/tmp/python-mutagen-20231008-6107-91zfxm/mutagen-1.47.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy @ file:///private/tmp/numpy-20240206-5996-spmzje/numpy-1.26.4
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyobjc==10.2
pyobjc-core==10.2
pyobjc-framework-Accessibility==10.2
pyobjc-framework-Accounts==10.2
pyobjc-framework-AddressBook==10.2
pyobjc-framework-AdServices==10.2
pyobjc-framework-AdSupport==10.2
pyobjc-framework-AppleScriptKit==10.2
pyobjc-framework-AppleScriptObjC==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-AppTrackingTransparency==10.2
pyobjc-framework-AudioVideoBridging==10.2
pyobjc-framework-AuthenticationServices==10.2
pyobjc-framework-AutomaticAssessmentConfiguration==10.2
pyobjc-framework-Automator==10.2
pyobjc-framework-AVFoundation==10.2
pyobjc-framework-AVKit==10.2
pyobjc-framework-AVRouting==10.2
pyobjc-framework-BackgroundAssets==10.2
pyobjc-framework-BusinessChat==10.2
pyobjc-framework-CalendarStore==10.2
pyobjc-framework-CallKit==10.2
pyobjc-framework-CFNetwork==10.2
pyobjc-framework-Cinematic==10.2
pyobjc-framework-ClassKit==10.2
pyobjc-framework-CloudKit==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Collaboration==10.2
pyobjc-framework-ColorSync==10.2
pyobjc-framework-Contacts==10.2
pyobjc-framework-ContactsUI==10.2
pyobjc-framework-CoreAudio==10.2
pyobjc-framework-CoreAudioKit==10.2
pyobjc-framework-CoreBluetooth==10.2
pyobjc-framework-CoreData==10.2
pyobjc-framework-CoreHaptics==10.2
pyobjc-framework-CoreLocation==10.2
pyobjc-framework-CoreMedia==10.2
pyobjc-framework-CoreMediaIO==10.2
pyobjc-framework-CoreMIDI==10.2
pyobjc-framework-CoreML==10.2
pyobjc-framework-CoreMotion==10.2
pyobjc-framework-CoreServices==10.2
pyobjc-framework-CoreSpotlight==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-CoreWLAN==10.2
pyobjc-framework-CryptoTokenKit==10.2
pyobjc-framework-DataDetection==10.2
pyobjc-framework-DeviceCheck==10.2
pyobjc-framework-DictionaryServices==10.2
pyobjc-framework-DiscRecording==10.2
pyobjc-framework-DiscRecordingUI==10.2
pyobjc-framework-DiskArbitration==10.2
pyobjc-framework-DVDPlayback==10.2
pyobjc-framework-EventKit==10.2
pyobjc-framework-ExceptionHandling==10.2
pyobjc-framework-ExecutionPolicy==10.2
pyobjc-framework-ExtensionKit==10.2
pyobjc-framework-ExternalAccessory==10.2
pyobjc-framework-FileProvider==10.2
pyobjc-framework-FileProviderUI==10.2
pyobjc-framework-FinderSync==10.2
pyobjc-framework-FSEvents==10.2
pyobjc-framework-GameCenter==10.2
pyobjc-framework-GameController==10.2
pyobjc-framework-GameKit==10.2
pyobjc-framework-GameplayKit==10.2
pyobjc-framework-HealthKit==10.2
pyobjc-framework-ImageCaptureCore==10.2
pyobjc-framework-InputMethodKit==10.2
pyobjc-framework-InstallerPlugins==10.2
pyobjc-framework-InstantMessage==10.2
pyobjc-framework-Intents==10.2
pyobjc-framework-IntentsUI==10.2
pyobjc-framework-IOBluetooth==10.2
pyobjc-framework-IOBluetoothUI==10.2
pyobjc-framework-IOSurface==10.2
pyobjc-framework-iTunesLibrary==10.2
pyobjc-framework-KernelManagement==10.2
pyobjc-framework-LatentSemanticMapping==10.2
pyobjc-framework-LaunchServices==10.2
pyobjc-framework-libdispatch==10.2
pyobjc-framework-libxpc==10.2
pyobjc-framework-LinkPresentation==10.2
pyobjc-framework-LocalAuthentication==10.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.2
pyobjc-framework-MailKit==10.2
pyobjc-framework-MapKit==10.2
pyobjc-framework-MediaAccessibility==10.2
pyobjc-framework-MediaLibrary==10.2
pyobjc-framework-MediaPlayer==10.2
pyobjc-framework-MediaToolbox==10.2
pyobjc-framework-Metal==10.2
pyobjc-framework-MetalFX==10.2
pyobjc-framework-MetalKit==10.2
pyobjc-framework-MetalPerformanceShaders==10.2
pyobjc-framework-MetalPerformanceShadersGraph==10.2
pyobjc-framework-MetricKit==10.2
pyobjc-framework-MLCompute==10.2
pyobjc-framework-ModelIO==10.2
pyobjc-framework-MultipeerConnectivity==10.2
pyobjc-framework-NaturalLanguage==10.2
pyobjc-framework-NetFS==10.2
pyobjc-framework-Network==10.2
pyobjc-framework-NetworkExtension==10.2
pyobjc-framework-NotificationCenter==10.2
pyobjc-framework-OpenDirectory==10.2
pyobjc-framework-OSAKit==10.2
pyobjc-framework-OSLog==10.2
pyobjc-framework-PassKit==10.2
pyobjc-framework-PencilKit==10.2
pyobjc-framework-PHASE==10.2
pyobjc-framework-Photos==10.2
pyobjc-framework-PhotosUI==10.2
pyobjc-framework-PreferencePanes==10.2
pyobjc-framework-PushKit==10.2
pyobjc-framework-Quartz==10.2
pyobjc-framework-QuickLookThumbnailing==10.2
pyobjc-framework-ReplayKit==10.2
pyobjc-framework-SafariServices==10.2
pyobjc-framework-SafetyKit==10.2
pyobjc-framework-SceneKit==10.2
pyobjc-framework-ScreenCaptureKit==10.2
pyobjc-framework-ScreenSaver==10.2
pyobjc-framework-ScreenTime==10.2
pyobjc-framework-ScriptingBridge==10.2
pyobjc-framework-SearchKit==10.2
pyobjc-framework-Security==10.2
pyobjc-framework-SecurityFoundation==10.2
pyobjc-framework-SecurityInterface==10.2
pyobjc-framework-SensitiveContentAnalysis==10.2
pyobjc-framework-ServiceManagement==10.2
pyobjc-framework-SharedWithYou==10.2
pyobjc-framework-SharedWithYouCore==10.2
pyobjc-framework-ShazamKit==10.2
pyobjc-framework-Social==10.2
pyobjc-framework-SoundAnalysis==10.2
pyobjc-framework-Speech==10.2
pyobjc-framework-SpriteKit==10.2
pyobjc-framework-StoreKit==10.2
pyobjc-framework-Symbols==10.2
pyobjc-framework-SyncServices==10.2
pyobjc-framework-SystemConfiguration==10.2
pyobjc-framework-SystemExtensions==10.2
pyobjc-framework-ThreadNetwork==10.2
pyobjc-framework-UniformTypeIdentifiers==10.2
pyobjc-framework-UserNotifications==10.2
pyobjc-framework-UserNotificationsUI==10.2
pyobjc-framework-VideoSubscriberAccount==10.2
pyobjc-framework-VideoToolbox==10.2
pyobjc-framework-Virtualization==10.2
pyobjc-framework-Vision==10.2
pyobjc-framework-WebKit==10.2
pyparsing==3.1.2
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-escpos==3.1
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
requests==2.31.0
rsa==4.9
simpleaudio==1.0.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
